+++ src/frontend/web_application/src/components/NewDraftForm/presenter.jsx
@@ -30,6 +30,7 @@ class NewDraftForm extends Component {
     onSend: PropTypes.func.isRequired,
     onChange: PropTypes.func,
     user: PropTypes.shape({}),
+    renderDraftMessageActionsContainer: PropTypes.func.isRequired,
     __: PropTypes.func.isRequired,
   };
   static defaultProps = {
@@ -39,7 +40,10 @@ class NewDraftForm extends Component {
     user: { contact: {} },
   };
 
-  state = { draft: { participants: [], subject: '', body: '' } };
+  state = {
+    draft: { participants: [], subject: '', body: '' },
+    hasChanged: false,
+  };
 
   componentWillMount() {
     this.setState(prevState => generateStateFromProps(this.props, prevState));
@@ -63,9 +67,15 @@ class NewDraftForm extends Component {
     const { name, value } = ev.target;
 
     this.setState((prevState) => {
-      const draft = { ...prevState.draft, [name]: value };
+      const draft = {
+        ...prevState.draft,
+        [name]: value,
+      };
 
-      return { draft };
+      return {
+        draft,
+        hasChanged: true,
+      };
     }, () => {
       this.props.onChange({ draft: this.state.draft });
     });
@@ -78,6 +88,7 @@ class NewDraftForm extends Component {
         // no need to merge author, backend does it
         participants: recipients,
       },
+      hasChanged: true,
     }), () => {
       this.props.onChange({ draft: this.state.draft });
     });
@@ -103,7 +114,7 @@ class NewDraftForm extends Component {
   }
 
   render() {
-    const { user, internalId, __ } = this.props;
+    const { user, internalId, renderDraftMessageActionsContainer, __ } = this.props;
     const dropdownId = uuidV1();
     const recipients = this.state.draft.participants && this.state.draft.participants
       .filter(participant => participant.type.toLowerCase() !== 'from');
@@ -132,7 +143,7 @@ class NewDraftForm extends Component {
               alignRight
               isMenu
               closeOnClick
-            />
+            >{renderDraftMessageActionsContainer()}</Dropdown>
           </TopRow>
           <BodyRow className="m-new-draft__body">
             <RecipientList
@@ -157,10 +168,23 @@ class NewDraftForm extends Component {
             />
           </BodyRow>
           <BottomRow className="m-new-draft__bottom-bar">
-            <Button className="m-new-draft__bottom-action" shape="plain" onClick={this.handleSend} icon="send" responsive="icon-only">
+            <Button
+              className="m-new-draft__bottom-action"
+              shape="plain"
+              onClick={this.handleSend}
+              icon="send"
+              responsive="icon-only"
+              disabled={!recipients || recipients.length === 0}
+            >
               {__('messages.compose.action.send')}
             </Button>
-            <Button className="m-new-draft__bottom-action" onClick={this.handleSave} icon="save" responsive="icon-only">
+            <Button
+              className="m-new-draft__bottom-action"
+              onClick={this.handleSave}
+              icon="save"
+              responsive="icon-only"
+              disabled={!this.state.hasChanged}
+            >
               {__('messages.compose.action.save')}
             </Button>
             <Button className="m-new-draft__bottom-action m-new-draft__bottom-action--editor" icon="editor" responsive="icon-only" />