+++ src/frontend/web_application/src/scenes/NewDraft/index.js
@@ -2,7 +2,10 @@ import { createSelector } from 'reselect';
 import { bindActionCreators, compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslator } from '@gandi/react-translate';
-import { editDraft, requestNewDraft, saveDraft, sendDraft } from '../../store/modules/draft-message';
+import { push } from 'react-router-redux';
+import { editDraft, requestNewDraft, saveDraft, sendDraft, clearDraft } from '../../store/modules/draft-message';
+import { withNotification } from '../../hoc/notification';
+import { deleteMessage } from '../../store/modules/message';
 import Presenter from './presenter';
 
 const messageDraftSelector = state => state.draftMessage.draftsByInternalId;
@@ -21,14 +24,30 @@ const mapStateToProps = createSelector(
     internalId,
   })
 );
-const mapDispatchToProps = dispatch => bindActionCreators({
+
+const onSaveDraft = ({ internalId, draft, message }, ownProps) => dispatch =>
+  dispatch(saveDraft({ internalId, draft, message }))
+    .then(() => {
+      const { __, notifySuccess } = ownProps;
+
+      return notifySuccess({ message: __('draft.feedback.saved') });
+    });
+
+const onDeleteMessage = ({ message, internalId }) => dispatch =>
+  dispatch(deleteMessage({ message }))
+    .then(() => dispatch(clearDraft({ internalId })))
+    .then(() => dispatch(push('/')));
+
+const mapDispatchToProps = (dispatch, ownProps) => bindActionCreators({
   requestNewDraft,
   editDraft,
-  saveDraft,
   sendDraft,
+  onSaveDraft: params => onSaveDraft(params, ownProps),
+  onDeleteMessage,
 }, dispatch);
 
 export default compose(
-  connect(mapStateToProps, mapDispatchToProps),
-  withTranslator()
+  withTranslator(),
+  withNotification(),
+  connect(mapStateToProps, mapDispatchToProps)
 )(Presenter);