+++ src/frontend/web_application/test/functional/features/22-compose-spec.js
@@ -21,9 +21,11 @@ describe('Compose new message', () => {
   describe('navigation between drafts', () => {
     it('creates a new draft', () => {
       const text1 = 'Compose creates a new draft';
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() => element.all(by.css('.m-navbar-item .m-item-link'))
@@ -55,9 +57,11 @@ describe('Compose new message', () => {
     it('composes a new message with no recipients', () => {
       const text1 = 'new message with no rcpts ';
 
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() =>
@@ -90,9 +94,11 @@ describe('Compose new message', () => {
     it('composes a new message with a known recipient', () => {
       const text1 = 'new message for a known recipient';
 
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() => {
@@ -128,9 +134,11 @@ describe('Compose new message', () => {
 
   describe('recipient search results manipulation', () => {
     it('has no suggestions for an already selected recipient', () => {
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() => {
@@ -169,10 +177,12 @@ describe('Compose new message', () => {
     });
 
     it('shows and hide rcpt search results', () => {
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
       const dropdownSelector = by.css('.m-recipient-list__search .m-dropdown');
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() => {
@@ -207,10 +217,12 @@ describe('Compose new message', () => {
     });
 
     it('can use keyboard arrows to select search result', () => {
-      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn--principal', __('Compose'));
+      const writeButtonSelector = by.cssContainingText('.m-call-to-action__btn', __('Compose'));
       const searchResultItemsSelector = by.css('.m-recipient-list__search-result');
 
       browser.get('/')
+      // XXX: click .btn--principal to force :hover callback actions
+        .then(() => element(by.css('.m-call-to-action__btn--principal')).click())
         .then(() => element(writeButtonSelector).click())
         .then(() => browser.wait(EC.presenceOf($('.m-new-draft')), 1000))
         .then(() => {